requirements.txt
requests
whois
dnspython
nmap
pdfkit
jinja2
rich
